requirements.txt
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mysql-connector-python==9.5.0
numpy==2.3.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
xmltodict==1.0.2
~                             